+++ src/main/java/de/neuland/pug4j/PugConfiguration.java
@@ -5,6 +5,7 @@ import de.neuland.pug4j.Pug4J.Mode;
 import de.neuland.pug4j.exceptions.PugCompilerException;
 import de.neuland.pug4j.exceptions.PugException;
 import de.neuland.pug4j.expression.ExpressionHandler;
+import de.neuland.pug4j.expression.GraalJsExpressionHandler;
 import de.neuland.pug4j.expression.JexlExpressionHandler;
 import de.neuland.pug4j.filter.CDATAFilter;
 import de.neuland.pug4j.filter.CssFilter;
@@ -37,7 +38,7 @@ public class PugConfiguration {
     private Map<String, Object> sharedVariables = new HashMap<String, Object>();
     private TemplateLoader templateLoader = new FileTemplateLoader();
     private ExpressionHandler expressionHandler = new JexlExpressionHandler();
-//    private ExpressionHandler expressionHandler = new GraalJsExpressionHandler();
+    //private ExpressionHandler expressionHandler = new GraalJsExpressionHandler();
     protected static final int MAX_ENTRIES = 1000;
 
     public PugConfiguration() {